+++ include/parser/heredoc.h
@@ -6,7 +6,7 @@
 /*   By: susumuyagi <susumuyagi@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/15 15:44:04 by susumuyagi        #+#    #+#             */
-/*   Updated: 2024/03/29 22:11:53 by susumuyagi       ###   ########.fr       */
+/*   Updated: 2024/04/08 11:51:50 by susumuyagi       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,6 +22,6 @@ void	free_heredoc(t_heredoc *heredoc);
 int		set_heredoc_delimiter(t_minishell *minish, t_token *tok);
 
 void	input_heredoc(t_minishell *minish);
-void	write_heredoc(t_minishell *minish, int idx);
+int		write_heredoc(t_minishell *minish, int idx);
 
 #endif
